fix(karma): simpler solution for karma-only code that does not leak
