+++ packages/@lwc/features/src/flags.ts
@@ -52,8 +52,14 @@ export function setFeatureFlag(name: FeatureFlagName, value: FeatureFlagValue): 
         );
         return;
     }
-    // This may seem redundant, but `process.env.NODE_ENV === 'test-karma-lwc'` is replaced by Karma tests
-    if (process.env.NODE_ENV === 'test-karma-lwc' || process.env.NODE_ENV !== 'production') {
+
+    // We want this code to run in both production and dev mode, hence Karma-only code here
+    /* --begin-karma-only-code--
+    flags[name] = value;
+    return;
+    --end-karma-only-code-- */
+
+    if (process.env.NODE_ENV !== 'production') {
         // Allow the same flag to be set more than once outside of production to enable testing
         flags[name] = value;
     } else {
@@ -75,8 +81,12 @@ export function setFeatureFlag(name: FeatureFlagName, value: FeatureFlagValue): 
  * purposes. It is a no-op when invoked in production mode.
  */
 export function setFeatureFlagForTest(name: FeatureFlagName, value: FeatureFlagValue): void {
-    // This may seem redundant, but `process.env.NODE_ENV === 'test-karma-lwc'` is replaced by Karma tests
-    if (process.env.NODE_ENV === 'test-karma-lwc' || process.env.NODE_ENV !== 'production') {
+    // We want this code to run in both production and dev mode, hence Karma-only code here
+    /* --begin-karma-only-code--
+    setFeatureFlag(name, value);
+    return;
+    --end-karma-only-code-- */
+    if (process.env.NODE_ENV !== 'production') {
         setFeatureFlag(name, value);
     }
 }
